moved booking data match assert helper
